Automatically Merged using SST Pull Request AutoTester

+++ src/sst/elements/firefly/nic.cc
@@ -29,7 +29,7 @@ using namespace SST::Firefly;
 using namespace SST::Interfaces;
 using namespace std::placeholders;
 
-int Nic::MaxPayload = (1L<<32) - 1;
+int Nic::MaxPayload = (int)((1L<<32) - 1);
 int Nic::m_packetId = 0;
 int Nic::ShmemSendMove::m_alignment = 64;
 
@@ -64,7 +64,6 @@ Nic::Nic(ComponentId_t id, Params &params) :
     m_txDelay =      params.find<int>( "txDelay_ns", 50 );
     int hostReadDelay = params.find<int>( "hostReadDelay_ns", 200 );
     m_shmemRxDelay_ns = params.find<int>( "shmemRxDelay_ns",0); 
-
 
     m_num_vNics = params.find<int>("num_vNics", 1 );
